fix: Update Node.js to 16 and package-lock to v2
The minimum Node.js version is now 16. The installation
script will use that from now on. Also the package-locks are now updated
to v2 which will probably make it not possible to run haxroomie with
earlier node versions.

Also fixed the gecos of the user that the installation script creates.
Add mention to readme that running install script with sudo will
probably fail.